requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
distro==1.9.0
Django==5.1.5
django-anymail==12.0
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
Markdown==3.7
openai==1.61.1
pillow==11.1.0
psycopg2==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
